tetragon/windows: Compile bpffs functions on Windows
While bppffs is not supported on Windows, we still need some stub functions to compile tetragon code.
This commit adds stub functions specfific to Windows compilation, which allow the bpf package to compile on Windows
